+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Code Editor Package for Visual Studio
 
+## [2.0.18] - 2023-03-17
+
+Integration:
+
+- Performance improvements with `EditorApplication.update` callbacks.
+ 
+Project generation:
+
+- Add extra compiler options for analyzers and source generators.
+
+
 ## [2.0.17] - 2022-12-06
 
 Integration: